fix: Filter by tags keeped when load more clicked.
