+++ src/hooks/stores/createPostStore.js
@@ -1,4 +1,6 @@
 export const createPostStore = (set) => ({
   postList: null,
+  tagId: null,
   setPostList: (_posts) => set(() => ({ postList: _posts })),
+  setTagId: (_tagId) => set(() => ({ tagId: _tagId })),
 });
